+++ app/Http/Controllers/HotelController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\NotifyUser;
 use App\Http\Requests\HotelRequest;
 use App\Http\Resources\HotelResource;
 use App\Models\Hotel;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
 
 class HotelController extends Controller
 {
@@ -22,6 +24,12 @@ class HotelController extends Controller
      */
     public function store(HotelRequest $request)
     {
+        $role = Role::findByName('admin', 'web');
+        $user = $request->user();
+        if (!($user->hasRole('admin') && $role->hasPermissionTo('create hotel'))) {
+            abort(403, 'You are not Authorized');
+        }
+
         $request->validated();
 
         Hotel::create([
@@ -31,6 +39,8 @@ class HotelController extends Controller
             "price_per_night" => $request->input('price_per_night'),
             "available_rooms" => $request->input('available_rooms')
         ]);
+
+        event(new NotifyUser(auth()->user(), 'Hotel Details Added'));
 
         return response(
             [
@@ -52,6 +62,12 @@ class HotelController extends Controller
      */
     public function update(Request $request, Hotel $hotel)
     {
+        $role = Role::findByName('admin', 'web');
+        $user = $request->user();
+        if (!($user->hasRole('admin') && $role->hasPermissionTo('update hotel'))) {
+            abort(403, 'You are not Authorized');
+        }
+
         $request->validate([
             'name' => 'min:4',
             'location' => 'min:4',
@@ -68,6 +84,8 @@ class HotelController extends Controller
             "available_rooms" => $request->available_rooms
         ]);
 
+        event(new NotifyUser(auth()->user(), "Details For $hotel->name has been updated"));
+
         return response(
             [
                 'message' => "$hotel->name Successfully Updated"
@@ -78,9 +96,17 @@ class HotelController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Hotel $hotel)
+    public function destroy(Hotel $hotel, Request $request)
     {
+        $role = Role::findByName('admin', 'web');
+        $user = $request->user();
+        if (!($user->hasRole('admin') && $role->hasPermissionTo('delete hotel'))) {
+            abort(403, 'You are not Authorized');
+        }
+
         $hotel->delete();
+
+        event(new NotifyUser(auth()->user(), "You have Successfull Deleted $hotel->name hotel record"));
 
         return response([
             'message' => 'Hotel Data Deleted'
